requirements.txt
# Python 3.12.10
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
distro==1.9.0
greenlet==3.2.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
jiter==0.9.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.25
langchain-core==0.3.58
langchain-mcp-adapters==0.0.10
langchain-openai==0.3.16
langchain-text-splitters==0.3.8
langsmith==0.3.42
mcp==1.7.1
openai==1.77.0
orjson==3.10.18
packaging==24.2
pydantic==2.11.4
pydantic-settings==2.9.1
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
sse-starlette==2.3.4
starlette==0.46.2
tenacity==9.1.2
tiktoken==0.9.0
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.0.7
uvicorn==0.34.2
worldnewsapi==2.2.0
zstandard==0.23.0

# Agent Angus dependencies - Real YouTube API integration
supabase>=2.0.0
google-auth>=2.17.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.88.0

# Additional dependencies for real tools
requests>=2.31.0
tempfile

# Yona dependencies (compatible versions)
fastapi>=0.110.0
uvicorn>=0.27.0

# Real Yona tools dependencies
httpx>=0.28.0
supabase>=2.0.0
